requirements.txt
Pillow
exif
pilgram